+++ NEWS.md
@@ -1,5 +1,6 @@
 # palaeoverse (development version)
 
+* Added age arguments (max_ma/min_ma) to bin_time (#106)
 * Added warning to bin_time (#104)
 * Improved link accessibility (#88)