+++ NEWS.md
@@ -1,8 +1,10 @@
 # palaeoverse (development version)
 
+* New function to create stratigraphic range plots, tax_range_strat (#97)
 * Added age arguments (max_ma/min_ma) to bin_time (#106)
 * Added warning to bin_time (#104)
 * Improved link accessibility (#88)
+
 
 # palaeoverse 1.2.1