+++ NEWS.md
@@ -1,11 +1,11 @@
 # palaeoverse (development version)
 
+* Update palaeorotate to handle updates with GPlates API (#112)
 * New function to create stratigraphic range plots, tax_range_strat (#97)
 * Added age arguments (max_ma/min_ma) to bin_time (#106)
 * Added warning to bin_time (#104)
 * Improved link accessibility (#88)
 * Fix superfluous columns in palaeorotate (#110)
-
 
 # palaeoverse 1.2.1